Add 'multi-word' alias for git command.
I frequently misspell the command 'git status' as 'git sttaus'.
Overeager I guess.  Anyway, this commit adds a bash function as a
wrapper around git.  If the first argument is 'sttaus', it invokes
git status.  If the argument is anything else, it just treats it as
a normal git command.